Fix modal spacing issue, add description to config form fields
