+++ new-log-viewer/src/components/MenuBar.tsx
@@ -4,29 +4,18 @@ import React, {
 } from "react";
 
 import {
-    Button,
-    DialogActions,
-    DialogContent,
-    DialogTitle,
     IconButton,
     Modal,
-    ModalDialog,
     Sheet,
-    ToggleButtonGroup,
     Typography,
-    useColorScheme,
 } from "@mui/joy";
-import type {Mode} from "@mui/system/cssVars/useCurrentColorScheme";
 
 import {SvgIconComponent} from "@mui/icons-material";
-import DarkMode from "@mui/icons-material/DarkMode";
 import Description from "@mui/icons-material/Description";
 import FileOpenIcon from "@mui/icons-material/FileOpen";
-import LightMode from "@mui/icons-material/LightMode";
 import NavigateBefore from "@mui/icons-material/NavigateBefore";
 import NavigateNext from "@mui/icons-material/NavigateNext";
 import Settings from "@mui/icons-material/Settings";
-import SettingsBrightnessIcon from "@mui/icons-material/SettingsBrightness";
 import SkipNext from "@mui/icons-material/SkipNext";
 import SkipPrevious from "@mui/icons-material/SkipPrevious";
 import TipsAndUpdates from "@mui/icons-material/TipsAndUpdates";
@@ -36,23 +25,16 @@ import {
     updateWindowUrlHashParams,
     UrlContext,
 } from "../contexts/UrlContextProvider";
-import {
-    CONFIG_KEY,
-    LOCAL_STORAGE_KEY,
-    THEME_NAME,
-} from "../typings/config";
+import {CONFIG_KEY} from "../typings/config";
 import {CURSOR_CODE} from "../typings/worker";
 import {ACTION_NAME} from "../utils/actions";
-import {
-    getConfig,
-    setConfig,
-} from "../utils/config";
+import {getConfig} from "../utils/config";
 import {openFile} from "../utils/file";
 import {
     getFirstItemNumInNextChunk,
     getLastItemNumInPrevChunk,
 } from "../utils/math";
-import ConfigDialog from "./ConfigDialog";
+import ConfigDialog from "./modals/SettingsModal/ConfigDialog";
 
 
 /**
@@ -85,64 +67,11 @@ export const handleAction = (actionName: ACTION_NAME, logEventNum: number, numEv
     }
 };
 
-/**
- * Handles the reset event for the configuration form.
- *
- * @param ev The form event triggered by the reset action.
- * @return
- */
-const handleConfigFormReset = (ev: React.FormEvent) => {
-    ev.preventDefault();
-    window.localStorage.clear();
-    window.location.reload();
-};
-
-/**
- * Handles the submit event for the configuration form.
- *
- * @param ev The form event triggered by the submit action.
- * @return
- */
-const handleConfigFormSubmit = (ev: React.FormEvent) => {
-    ev.preventDefault();
-    const formData = new FormData(ev.target as HTMLFormElement);
-
-    const formatString = formData.get(LOCAL_STORAGE_KEY.DECODER_OPTIONS_FORMAT_STRING);
-    const logLevelKey = formData.get(LOCAL_STORAGE_KEY.DECODER_OPTIONS_LOG_LEVEL_KEY);
-    const timestampKey = formData.get(LOCAL_STORAGE_KEY.DECODER_OPTIONS_TIMESTAMP_KEY);
-    const pageSize = formData.get(LOCAL_STORAGE_KEY.PAGE_SIZE);
-    let error = null;
-    if (
-        "string" === typeof formatString &&
-        "string" === typeof logLevelKey &&
-        "string" === typeof timestampKey
-    ) {
-        error ||= setConfig({
-            key: CONFIG_KEY.DECODER_OPTIONS,
-            value: {formatString, logLevelKey, timestampKey},
-        });
-    }
-    if ("string" === typeof pageSize) {
-        error ||= setConfig({
-            key: CONFIG_KEY.PAGE_SIZE,
-            value: Number(pageSize),
-        });
-    }
-    if (null !== error) {
-        // eslint-disable-next-line no-warning-comments
-        // TODO: Show an error pop-up once NotificationProvider is implemented.
-        // eslint-disable-next-line no-alert
-        window.alert(error);
-    } else {
-        window.location.reload();
-    }
-};
 
 /**
  *
  */
 export const MenuBar = () => {
-    const {setMode, mode} = useColorScheme();
     const {logEventNum} = useContext(UrlContext);
     const {fileName, loadFile, numEvents} = useContext(StateContext);
 
@@ -226,61 +155,7 @@ export const MenuBar = () => {
                     setSettingsModelOpen(false);
                 }}
             >
-                <form
-                    onReset={handleConfigFormReset}
-                    onSubmit={handleConfigFormSubmit}
-                >
-                    <ModalDialog layout={"fullscreen"}>
-                        <DialogTitle className={"menu-bar-modal"}>
-                            <span style={{flexGrow: 1}}>
-                                Settings
-                            </span>
-                            <ToggleButtonGroup
-                                size={"sm"}
-                                value={mode as string}
-                                onChange={(_, newValue) => {
-                                    setMode(newValue as Mode);
-                                }}
-                            >
-                                <Button
-                                    startDecorator={<LightMode/>}
-                                    value={THEME_NAME.LIGHT}
-                                >
-                                    Light
-                                </Button>
-                                <Button
-                                    startDecorator={<SettingsBrightnessIcon/>}
-                                    value={THEME_NAME.SYSTEM}
-                                >
-                                    System
-                                </Button>
-                                <Button
-                                    startDecorator={<DarkMode/>}
-                                    value={THEME_NAME.DARK}
-                                >
-                                    Dark
-                                </Button>
-                            </ToggleButtonGroup>
-                        </DialogTitle>
-                        <DialogContent>
-                            <ConfigDialog/>
-                        </DialogContent>
-                        <DialogActions>
-                            <Button
-                                color={"primary"}
-                                type={"submit"}
-                            >
-                                Apply & Reload
-                            </Button>
-                            <Button
-                                color={"neutral"}
-                                type={"reset"}
-                            >
-                                Reset Default
-                            </Button>
-                        </DialogActions>
-                    </ModalDialog>
-                </form>
+                <ConfigDialog/>
             </Modal>
         </Sheet>
     );
